+++ lib/i18n/strings_en_us.dart
@@ -17,18 +17,18 @@ class _StringsEnUs extends _StringsI18n {
   @override
   List<String> getMonths() {
     return [
-      "January",
-      "February",
-      "March",
-      "April",
+      "Jan",
+      "Feb",
+      "Mar",
+      "Apr",
       "May",
-      "June",
-      "July",
-      "August",
-      "September",
-      "October",
-      "November",
-      "December"
+      "Jun",
+      "Jul",
+      "Aug",
+      "Sep",
+      "Oct",
+      "Nov",
+      "Dec"
     ];
   }