+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.1]
+
+* Modified diameter ratio to make datepicker flat
+* reduce thickness
+
 ## [0.1.0]
 
 * Initial release.